gallery config

+++ src/main/java/crocobob/CIENderella/repository/GalleryInfo/GalleryInfoRepository_SpringDataJpa.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GalleryInfoRepository_SpringDataJpa extends JpaRepository<GalleryInfo, Integer>, GalleryInfoRepository {
+public interface GalleryInfoRepository_SpringDataJpa extends JpaRepository<GalleryInfo, Long>, GalleryInfoRepository {
 }
